refactor(Calculator): 과제 스텝 변경에 다른 패키지 추가
기존 과제를 step1 패키지로 이동
