+++ src/main/java/racingcar/step1/InputUI.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.step1;
 
 import java.util.List;
 import java.util.Scanner;
