+++ src/main/java/racingcar/step1/OutputUI.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.step1;
 
 import java.util.List;
